+++ tests/plugins/test_ticking.py
@@ -1,0 +1,87 @@
+import time
+from datetime import datetime as dt
+from datetime import timedelta
+from os.path import dirname, join
+from typing import Callable
+from unittest.mock import patch
+
+import gi
+import pytest
+
+from tomate.pomodoro import SessionPayload, SessionType
+
+gi.require_version("Gtk", "3.0")
+gi.require_version("Gst", "1.0")
+
+from tomate.pomodoro import Events
+
+DEFAULT_ALARM = f'file://{join(dirname(dirname(__file__)), "data", "tomate", "media", "clock.ogg")}'
+
+
+def wait_until(fn: Callable[[], bool], timeout: int = 1, period: int = 0.25):
+    limit = dt.utcnow() + timedelta(seconds=timeout)
+
+    while dt.utcnow() < limit:
+        if fn():
+            return True
+        time.sleep(period)
+    return False
+
+
+@pytest.fixture
+def plugin(bus, config, graph):
+    graph.providers.clear()
+    graph.register_instance("tomate.config", config)
+    graph.register_instance("tomate.bus", bus)
+
+    from ticking import TickingPlugin
+
+    instance = TickingPlugin()
+    instance.configure(bus, graph)
+    return instance
+
+
+class TestPlugin:
+    def test_loads_plugin_configuration(self, bus, config, plugin):
+        plugin.activate()
+
+        assert plugin.player.volume == 0.5
+        assert plugin.player.file == DEFAULT_ALARM
+        assert plugin.player.repeat
+
+    @patch("ticking.GStreamerPlayer")
+    def test_starts_player_when_session_start(self, player, bus, config, plugin):
+        plugin.activate()
+
+        bus.send(Events.SESSION_START, payload=create_session_payload())
+
+        player.return_value.play.assert_called_once()
+
+    @patch("ticking.GStreamerPlayer")
+    @pytest.mark.parametrize("event", [Events.SESSION_END, Events.SESSION_INTERRUPT])
+    def test_stops_player_when_session_(self, player, event, bus, config, plugin):
+        plugin.activate()
+
+        bus.send(Events.SESSION_START, payload=create_session_payload())
+        bus.send(event)
+
+        player.return_value.stop.assert_called_once()
+
+    @patch("ticking.GStreamerPlayer")
+    def test_stops_player_when_plugin_deactivate(self, player, bus, config, plugin):
+        plugin.activate()
+
+        plugin.deactivate()
+
+        player.return_value.stop.assert_called_once()
+
+
+def create_session_payload(**kwargs) -> SessionPayload:
+    defaults = {
+        "id": "1234",
+        "duration": 25 * 60,
+        "pomodoros": 0,
+        "type": SessionType.POMODORO,
+    }
+    defaults.update(kwargs)
+    return SessionPayload(**defaults)